health_x/03_automation/requirements.txt
anthropic>=0.40.0
tweepy>=4.14.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
